refactor: move filter modules under `filter` packages
