+++ src/test/java/gdsc/konkuk/platformcore/controller/member/MemberControllerTest.java
@@ -22,7 +22,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import com.epages.restdocs.apispec.ResourceSnippetParameters;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import gdsc.konkuk.platformcore.util.annotation.RestDocsTest;
-import gdsc.konkuk.platformcore.application.auth.JwtTokenProvider;
+import gdsc.konkuk.platformcore.filter.auth.JwtTokenProvider;
 import gdsc.konkuk.platformcore.application.member.MemberService;
 import gdsc.konkuk.platformcore.application.member.dtos.MemberCreateCommand;
 import gdsc.konkuk.platformcore.application.member.exceptions.UserAlreadyExistException;
